finishing out Android functionality
