+++ Sensus.Shared/Notifications/INotifyMeta.cs
@@ -33,53 +33,4 @@ namespace Sensus.Shared.Notifications
         /// </summary>
         bool LagAllowed { get; set; } //public const string SENSUS_CALLBACK_REPEAT_LAG_KEY = "SENSUS-CALLBACK-REPEAT-LAG";
     }
-
-    public static class INotifyMetaExtensions
-    {
-        public static void Execute(this INotifyMeta meta)
-        {
-            var serviceHelper  = SensusServiceHelper.Get();
-            var scheduleHelper = (CallbackScheduler)SensusContext.Current.CallbackScheduler;
-
-            if (meta == null)
-            {
-                serviceHelper.LetDeviceSleep();
-                return;
-            }
-
-            if (meta.Type == NotificationType.Undefined)
-            {
-                serviceHelper.LetDeviceSleep();
-                return;
-            }
-
-            if (meta.Type == NotificationType.Callback && !scheduleHelper.CallbackIsScheduled(meta.CallbackId))
-            {
-                scheduleHelper.UnscheduleCallback(meta.CallbackId);
-                serviceHelper.LetDeviceSleep();
-            }
-
-            if(meta.Type == NotificationType.Callback && scheduleHelper.CallbackIsScheduled(meta.CallbackId))
-            { 
-            }
-
-            if (meta.Type == NotificationType.Script)
-            {
-                serviceHelper.BringToForeground();
-
-                // display the pending scripts page if it is not already on the top of the navigation stack
-                SensusContext.Current.MainThreadSynchronizer.ExecuteThreadSafe(async () =>
-                {
-                    var topPage = Xamarin.Forms.Application.Current.MainPage.Navigation.NavigationStack.LastOrDefault();
-
-                    if (!(topPage is PendingScriptsPage))
-                    {
-                        await Xamarin.Forms.Application.Current.MainPage.Navigation.PushAsync(new PendingScriptsPage());
-                    }
-
-                    serviceHelper.LetDeviceSleep();
-                });
-            }
-        }
-    }
 }
